aqi_prediction/requirements.txt
boto3>=1.26.0
pandas>=1.5.3
numpy>=1.24.0
requests>=2.28.0
scikit-learn>=1.2.0
autogluon>=1.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
plotly>=5.15.0
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0 